scrpto: Move KeyMeta + Version
Previously, the TRC package defined its own KeyMeta and Version format.
However, this will be reused by the issuer and AS certificates.

This commit moves the definitions to scrypto.
