+++ go/lib/scrypto/trc/v2/signed_test.go
@@ -37,7 +37,7 @@ func TestEncode(t *testing.T) {
 		},
 		"Invalid Version": {
 			Modify: func(base *trc.TRC) {
-				base.Version = trc.Version(scrypto.LatestVer)
+				base.Version = scrypto.Version(scrypto.LatestVer)
 			},
 			Assertion: assert.Error,
 		},
